Hentai's config @ 15th place :-) More detail below:
This is NOT what he had on the car, I used emstudio fb7790e9ad7c0c476ea08ecf9b3e914d88a1e71e to pull the following set of files out of a fresh build of his base code and his last S19 which he emailed me:

http://stuff.fredcooke.com/Hentai.tune.base.and.final.zip

Additionally, the decoder and build config is removed, as it was not ready for publication and was pending his test results.

Finally, his IAT is hard coded in this setup. A real setup would change that config and run a sensor...

Hopefully he honours his word and helps us test and improve the system.